requirements.txt
textarena
transformers>=4.35.0
torch>=2.0.0
gradio>=4.0.0
flask>=2.0.0
numpy>=1.20.0
tqdm>=4.64.0
streamlit
openai>=1.0.0
PyYAML>=6.0